refactor: landing template
