+++ interfaces/index.ts
@@ -1,13 +1,6 @@
-export enum UserType {
-  normalUser,
-  admin,
-  superAdmin
-}
-
 export type User = {
-  id: string;
-  username: string;
-  password: string;
+  name: string;
   email: string;
-  type: UserType;
+  role: string;
+  imageUrl: string;
 };
