+++ pages/_app.tsx
@@ -1,9 +1,14 @@
-import '../styles/globals.css'
+import '../styles/globals.css';
 import 'tailwindcss/tailwind.css';
-import type { AppProps } from 'next/app'
+import type { AppProps } from 'next/app';
+import { MantineProvider } from '@mantine/core';
 
-function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
+const MyApp = ({ Component, pageProps }: AppProps) => {
+  return (
+    <MantineProvider withGlobalStyles withNormalizeCSS>
+      <Component {...pageProps} />
+    </MantineProvider>
+  );
 }
 
-export default MyApp
+export default MyApp;
