+++ utils/classNames.ts
@@ -1,0 +1,4 @@
+const classNames = (...classes: string[]) => {
+  return classes.filter(Boolean).join(' ');
+};
+export default classNames;
